apps/api/requirements.txt
# Keep in sync with `environment.yml` (pip section).
fastapi==0.128.5
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
alembic==1.18.3
psycopg[binary]==3.3.2
pydantic==2.12.5
email-validator==2.3.0
authlib==1.6.7
itsdangerous==2.2.0
httpx==0.28.1
python-dotenv==1.2.1
bcrypt==5.0.0
langgraph==1.0.8

# Dev/test tooling
pytest==9.0.2
ruff==0.15.0
